Rename "dash" package to "grizzly"
