+++ pkg/grizzly/config_test.go
@@ -1,4 +1,4 @@
-package dash
+package grizzly
 
 import (
 	"os"
