+++ pkg/grizzly/jsonnet.go
@@ -1,4 +1,4 @@
-package dash
+package grizzly
 
 import (
 	"github.com/google/go-jsonnet"
